sulambi-backend-main/requirements.txt
flask
flask-cors
python-dotenv
scikit-learn